Vulnerability: Move CVSS2 and CVSS3 rating enums to VulnerabilityReference
As the severity is now stored in VulnerabilityReference, it also makes more
sense to move the enums that map severity scores to ratings to that class.
