+++ model/src/main/kotlin/VulnerabilityReference.kt
@@ -49,4 +49,56 @@ data class VulnerabilityReference(
      * severity.
      */
     val severity: String?
-)
+) {
+    /**
+     * The rating attaches human-readable semantics to the score number according to CVSS version 2, see
+     * https://www.balbix.com/insights/cvss-v2-vs-cvss-v3/#CVSSv3-Scoring-Scale-vs-CVSSv2-6.
+     */
+    enum class Cvss2Rating(private val upperBound: Float) {
+        LOW(4.0f),
+        MEDIUM(7.0f),
+        HIGH(10.0f);
+
+        companion object {
+            /**
+             * Get the [Cvss2Rating] from a [score], or null if the [score] does not map to any [Cvss2Rating].
+             */
+            fun fromScore(score: Float): Cvss2Rating? =
+                when {
+                    score < 0.0f || score > HIGH.upperBound -> null
+                    score < LOW.upperBound -> LOW
+                    score < MEDIUM.upperBound -> MEDIUM
+                    score <= HIGH.upperBound -> HIGH
+                    else -> null
+                }
+        }
+    }
+
+    /**
+     * The rating attaches human-readable semantics to the score number according to CVSS version 3, see
+     * https://www.first.org/cvss/v3.0/specification-document#Qualitative-Severity-Rating-Scale.
+     */
+    enum class Cvss3Rating(private val upperBound: Float) {
+        NONE(0.0f),
+        LOW(4.0f),
+        MEDIUM(7.0f),
+        HIGH(9.0f),
+        CRITICAL(10.0f);
+
+        companion object {
+            /**
+             * Get the [Cvss3Rating] from a [score], or null if the [score] does not map to any [Cvss3Rating].
+             */
+            fun fromScore(score: Float): Cvss3Rating? =
+                when {
+                    score < 0.0f || score > CRITICAL.upperBound -> null
+                    score == NONE.upperBound -> NONE
+                    score < LOW.upperBound -> LOW
+                    score < MEDIUM.upperBound -> MEDIUM
+                    score < HIGH.upperBound -> HIGH
+                    score <= CRITICAL.upperBound -> CRITICAL
+                    else -> null
+                }
+        }
+    }
+}
