+++ model/src/test/kotlin/VulnerabilityTest.kt
@@ -27,8 +27,8 @@ import io.kotest.matchers.shouldBe
 
 import java.lang.IllegalArgumentException
 
-import org.ossreviewtoolkit.model.Vulnerability.Cvss2Rating
-import org.ossreviewtoolkit.model.Vulnerability.Cvss3Rating
+import org.ossreviewtoolkit.model.VulnerabilityReference.Cvss2Rating
+import org.ossreviewtoolkit.model.VulnerabilityReference.Cvss3Rating
 
 class VulnerabilityTest : StringSpec({
     "The CVSS 2 rating should be correct for a given score" {
